requirements.txt
matplotlib==3.3.4
numpy==1.20.1
Pillow==9.1.0
streamlit==1.8.1
tensorflow==2.8.0
tensorflow_hub==0.12.0
~treamlit==1.8.0
